VCV-224: Remove unused files
