--- src/features/review-legacy/dhis2-sync/utils/dhis2-sync-site-status.ts
@@ -1,9 +1,0 @@
-export type Dhis2SyncSiteStatus =
-    | 'ready'
-    | 'queued'
-    | 'running'
-    | 'submitted'
-    | 'failed'
-    | 'timedOut'
-    | 'skipped'
-    | 'hasNewCertifiedData';
